RavenDB-18689 Sharding - indexes view - enhacements
