+++ src/Raven.Studio/typescript/components/pages/database/indexes/list/IndexesPage.spec.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import { composeStories } from "@storybook/testing-react";
 
 import * as stories from "./IndexesPage.stories";
-const { EmptyView, SampleDataCluster } = composeStories(stories);
+const { EmptyView, SampleDataCluster, FaultyIndexSharded, FaultyIndexSingleNode } = composeStories(stories);
 
 function render() {
     const db = DatabasesStubs.shardedDatabase();
@@ -30,5 +30,33 @@ describe("IndexesPage", function () {
         await screen.findByText("ReplacementOf/Orders/ByCompany");
         const deleteButtons = await screen.findAllByTitle(/Delete the index/i);
         expect(deleteButtons.length).toBeGreaterThanOrEqual(1);
+    });
+
+    it("can show search engine - corax", async () => {
+        const { screen, getQueriesForElement } = rtlRender(<SampleDataCluster />);
+
+        const orderTotals = await screen.findByText("Orders/ByCompany");
+        const indexItem = orderTotals.closest(".rich-panel-item");
+        const indexItemSelectors = getQueriesForElement(indexItem);
+
+        expect(await indexItemSelectors.findByText(/Corax/)).toBeInTheDocument();
+    });
+
+    it("can open faulty index - sharded", async () => {
+        const { screen } = rtlRender(<FaultyIndexSharded />);
+
+        const openFaultyButton = await screen.findByText(/Open faulty index/);
+        expect(openFaultyButton).toBeInTheDocument();
+
+        expect(screen.queryByText(/Set State/)).not.toBeInTheDocument();
+    });
+
+    it("can open faulty index - single node", async () => {
+        const { screen } = rtlRender(<FaultyIndexSingleNode />);
+
+        const openFaultyButton = await screen.findByText(/Open faulty index/);
+        expect(openFaultyButton).toBeInTheDocument();
+
+        expect(screen.queryByText(/Set State/)).not.toBeInTheDocument();
     });
 });
